requirements.txt
langchain==0.2.1
langchain-anthropic==0.1.15
langchain-community==0.2.1
langchain-core==0.2.3
langchain-elasticsearch==0.2.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langchainhub==0.1.17
langgraph==0.0.60
langsmith==0.1.67
openai==1.30.5
PyGithub==2.3.0
kubernetes==29.0.0
